Tests/LibWeb/Text/expected/css/animated-background-image-timer-count-is-document-scoped.txt

top document has animation state: false
iframe document timer active: true
